Moved open terminal button to inside misc settings
Was too dangerous to have next to minimize
