+++ Wabbajack/View Models/Settings/SettingsVM.cs
@@ -1,8 +1,12 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Diagnostics;
+using System.IO;
 using System.Linq;
+using System.Reflection;
 using System.Text;
 using System.Threading.Tasks;
+using System.Windows.Input;
 using ReactiveUI;
 using Wabbajack.Lib;
 
@@ -15,7 +19,9 @@ namespace Wabbajack
         public PerformanceSettings Performance { get; }
         public FiltersSettings Filters { get; }
         public AuthorFilesVM AuthorFile { get; }
-        
+
+        public ICommand OpenTerminalCommand { get; }
+
         public SettingsVM(MainWindowVM mainWindowVM)
             : base(mainWindowVM)
         {
@@ -24,7 +30,18 @@ namespace Wabbajack
             Performance = mainWindowVM.Settings.Performance;
             AuthorFile = new AuthorFilesVM(this);
             Filters = mainWindowVM.Settings.Filters;
+            OpenTerminalCommand = ReactiveCommand.CreateFromTask(() => OpenTerminal());
         }
 
+        private async Task OpenTerminal()
+        {
+            var process = new ProcessStartInfo
+            {
+                FileName = "cmd.exe",
+                WorkingDirectory = Path.GetDirectoryName(Assembly.GetEntryAssembly().Location)
+            };
+            Process.Start(process);
+            await MWVM.ShutdownApplication();
+        }
     }
 }
