+++ Wabbajack/Views/MainWindow.xaml
@@ -69,14 +69,6 @@
                     Height="17"
                     Kind="Cog" />
             </Button>
-            <Button Grid.Column="1"
-                Margin="5,0"
-                Command="{Binding OpenTerminalCommand}">
-                <icon:PackIconMaterial
-                    Width="17"
-                    Height="17"
-                    Kind="Console" />
-            </Button>
         </mahapps:WindowCommands>
     </mahapps:MetroWindow.RightWindowCommands>
 </mahapps:MetroWindow>
