+++ Wabbajack/Views/Settings/MiscSettingsView.xaml
@@ -9,7 +9,7 @@
     xmlns:xwpf="http://schemas.xceed.com/wpf/xaml/toolkit"
     d:DesignHeight="450"
     d:DesignWidth="800"
-    x:TypeArguments="local:FiltersSettings"
+    x:TypeArguments="local:SettingsVM"
     mc:Ignorable="d">
     <Border
         x:Name="PerformanceView"
@@ -42,6 +42,7 @@
                     <RowDefinition />
                     <RowDefinition />
                     <RowDefinition />
+                    <RowDefinition />
                 </Grid.RowDefinitions>
                 <Grid.Resources>
                     <Style BasedOn="{StaticResource MainButtonStyle}" TargetType="Button">
@@ -55,21 +56,21 @@
                     </Style>
                 </Grid.Resources>
                 <CheckBox Grid.Row="0"
-                          Name="FilterPersistCheckBox"
-                          Margin="0,5,0,0"
-                          HorizontalAlignment="Left"
-                          VerticalAlignment="Top"
-                          Content="Gallery filters are saved on exit">
+                    Name="FilterPersistCheckBox"
+                    Margin="0,5,0,0"
+                    HorizontalAlignment="Left"
+                    VerticalAlignment="Top"
+                    Content="Gallery filters are saved on exit">
                     <CheckBox.LayoutTransform>
                         <ScaleTransform ScaleX="1.2" ScaleY="1.2" />
                     </CheckBox.LayoutTransform>
                 </CheckBox>
                 <CheckBox Grid.Row="1"
-                          Name="UseCompressionCheckBox"
-                          Margin="0,5,0,0"
-                          HorizontalAlignment="Left"
-                          VerticalAlignment="Top"
-                          Content="Use NTFS LZS compression during install">
+                    Name="UseCompressionCheckBox"
+                    Margin="0,5,0,0"
+                    HorizontalAlignment="Left"
+                    VerticalAlignment="Top"
+                    Content="Use NTFS LZS compression during install">
                     <CheckBox.LayoutTransform>
                         <ScaleTransform ScaleX="1.2" ScaleY="1.2" />
                     </CheckBox.LayoutTransform>
@@ -78,6 +79,10 @@
                     Name="ClearCefCache"
                     Margin="0,5,0,0"
                     Content="Clear In-App Browser Cache" />
+                <Button Grid.Row="3"
+                    Name="OpenTerminal"
+                    Margin="0,5,0,0"
+                    Content="Open Terminal and Close WJ" />
             </Grid>
         </Grid>
     </Border>
